// apps/web/generated/api/models/formTemplateStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type FormTemplateStatus = typeof FormTemplateStatus[keyof typeof FormTemplateStatus];


export const FormTemplateStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;