// apps/web/generated/api/models/projectPriority.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type ProjectPriority = typeof ProjectPriority[keyof typeof ProjectPriority];


export const ProjectPriority = {
  low: 'low',
  normal: 'normal',
  high: 'high',
  urgent: 'urgent',
} as const;