// apps/web/generated/api/models/projectVisibility.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type ProjectVisibility = typeof ProjectVisibility[keyof typeof ProjectVisibility];


export const ProjectVisibility = {
  public: 'public',
  private: 'private',
  invite_only: 'invite_only',
} as const;