// apps/web/generated/api/models/categoryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryResponse {
  category_id: string;
  category_key: string;
  name: string;
  slug: string;
  description: string | null;
  is_active: boolean;
  sort_order: number;
  parent_category_id: string | null;
}