// apps/web/generated/api/models/formTemplateResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FormFieldResponse } from './formFieldResponse';
import type { FormTemplateStatus } from './formTemplateStatus';

export interface FormTemplateResponse {
  template_id: string;
  category_id: string;
  template_key: string;
  name: string;
  version_no: number;
  status: FormTemplateStatus;
  is_active: boolean;
  published_at: string | null;
  fields: FormFieldResponse[];
}